Update version number and date   modified:   storage/connect/ha_connect.cc
Add conditional SE exception support
  modified:   storage/connect/json.cpp
  modified:   storage/connect/plgdbutl.cpp

Change %p in %x in some sprintf functions.
This to avoid some compiler warnings.
  modified:   storage/connect/tabwmi.cpp
  modified:   storage/connect/tabxml.cpp
  modified:   storage/connect/value.h

Add JavaWrappers.jar to the class path
  modified:   storage/connect/jdbconn.cpp

Fix wrong declare (char *buf[256]; --> char  buf[256];)
  modified:   storage/connect/xindex.cpp

+++ storage/connect/value.h
@@ -271,7 +271,7 @@ class DllExport TYPVAL<PSZ>: public VALUE {
   virtual void   Reset(void) {*Strp = 0;}
   virtual int    GetValLen(void) {return Len;};
   virtual int    GetValPrec() {return (Ci) ? 1 : 0;}
-  virtual int    GetSize(void) {return (Strp) ? strlen(Strp) : 0;}
+  virtual int    GetSize(void) {return (Strp) ? (int)strlen(Strp) : 0;}
   virtual PSZ    GetCharValue(void) {return Strp;}
   virtual char   GetTinyValue(void);
   virtual uchar  GetUTinyValue(void);
